requirements.txt
python-telegram-bot==20.4
pytz==2023.3
APScheduler==3.10.1
